systeme/OCR_tamil_char/requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.8.0  # For image processing and character segmentation

# Progress Bars and Utilities
tqdm>=4.65.0

# Visualization and Evaluation
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0

# Type Hints
typing-extensions>=4.5.0  # For Python type hints compatibility 